Updating RileyLink frameworks

+++ Loop/Extensions/PumpModel.swift
@@ -14,7 +14,7 @@ extension PumpModel {
         return Int(rawValue)! / 100
     }
     
-    var reservoirCapacity: Double {
+    var reservoirCapacity: Int {
         switch size {
         case 5:
             return 176
